requirements.txt
appnope==0.1.2
asttokens==2.0.5
backcall==0.2.0
Brotli==1.0.9
cachetools==5.0.0
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.0.4
dash==2.3.0
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
debugpy==1.5.1
decorator==5.1.1
entrypoints==0.4
executing==0.8.3
Flask==2.0.3
Flask-Compress==1.11
google-api-core==2.7.1
google-auth==2.6.2
google-cloud==0.34.0
google-cloud-bigquery==2.34.2
google-cloud-core==2.2.3
google-cloud-storage==2.2.1
google-crc32c==1.3.0
google-resumable-media==2.3.2
googleapis-common-protos==1.56.0
grpcio==1.44.0
grpcio-status==1.44.0
idna==3.3
ipykernel==6.9.2
ipython==8.1.1
itsdangerous==2.1.1
jedi==0.18.1
Jinja2==3.0.3
jupyter-client==7.1.2
jupyter-core==4.9.2
MarkupSafe==2.1.1
matplotlib-inline==0.1.3
nest-asyncio==1.5.4
numpy==1.22.3
packaging==21.3
pandas==1.4.1
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
plotly==5.6.0
prompt-toolkit==3.0.28
proto-plus==1.20.3
protobuf==3.19.4
psutil==5.9.0
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==7.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
Pygments==2.11.2
pyparsing==3.0.7
python-dateutil==2.8.2
pytz==2021.3
pyzmq==22.3.0
requests==2.27.1
rsa==4.8
six==1.16.0
stack-data==0.2.0
tenacity==8.0.1
tornado==6.1
traitlets==5.1.1
urllib3==1.26.9
wcwidth==0.2.5
Werkzeug==2.0.3
